Database bad request exception

+++ src/main/java/br/com/mcoder/ecommerce/services/ProductService.java
@@ -3,11 +3,15 @@ package br.com.mcoder.ecommerce.services;
 import br.com.mcoder.ecommerce.dto.ProductDTO;
 import br.com.mcoder.ecommerce.entities.Product;
 import br.com.mcoder.ecommerce.repositories.ProductRepository;
+import br.com.mcoder.ecommerce.services.exceptions.DatabaseException;
 import br.com.mcoder.ecommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -43,14 +47,26 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO productDTO) {
-        Product product = repository.getReferenceById(id);
-        copyDtoEntity(productDTO, product);
-        product = repository.save(product);
-        return new ProductDTO(product);
+        try {
+            Product product = repository.getReferenceById(id);
+            copyDtoEntity(productDTO, product);
+            product = repository.save(product);
+            return new ProductDTO(product);
+        }catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
     private void copyDtoEntity(ProductDTO productDTO, Product product) {
         product.setName(productDTO.getName());
